Add CMake and GCC support
