+++ PrimeFDG.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27004.2005
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PrimeFDG", "PrimeFDG\PrimeFDG.vcxproj", "{5CDA2CB3-35DD-4C0D-B0E9-1AF02E49CF31}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PrimeFDG", "src\PrimeFDG.vcxproj", "{5CDA2CB3-35DD-4C0D-B0E9-1AF02E49CF31}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
