--- PrimeFDG/stdafx.h
@@ -1,19 +1,0 @@
-﻿// stdafx.h : include file for standard system include files,
-// or project specific include files that are used frequently, but
-// are changed infrequently
-//
-
-#pragma once
-
-#include "targetver.h"
-
-#include <stdio.h>
-#include <tchar.h>
-
-#include <stdint.h>
-#include <stdlib.h>
-#include <omp.h>
-#include <time.h>
-
-#include "options.h"
-#include "pfmath.h"
