+++ src/cpuid.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "options.h"
 
 extern bool cpuid_popcnt;
