+++ src/pattern.h
@@ -9,7 +9,7 @@
 // The prefix of the precomputed pattern
 static const BITARRAY_WORD pfdg_pattern_prefix = 0x7E92ED659B4B3491;
 // The precomputed pattern data
-static const __declspec(align(32)) BITARRAY_WORD pfdg_pattern[PFDG_PATTERN_LENGTH] =
+static const ALIGN32 BITARRAY_WORD pfdg_pattern[PFDG_PATTERN_LENGTH] =
 {
 	0xD6693DF279A5B3CD,
 	0x4B669EDDA5DACB36,
