+++ src/targetver.h
@@ -1,5 +1,6 @@
 ﻿#pragma once
 
+#ifdef _WIN32
 // Including SDKDDKVer.h defines the highest available Windows platform.
 
 // If you wish to build your application for a previous Windows platform, include WinSDKVer.h and
@@ -8,3 +9,5 @@
 #include <WinSDKVer.h>
 #define _WIN32_WINNT 0x0601
 #include <SDKDDKVer.h>
+
+#endif
